requirements.txt
gradio
pillow
flask